Create testcord.py
Added new module. `testcord.py` is for testing out an asychio.time() function to have two chatbots running on the same LLM at the same time, responding to eachother without getting hung on the wait when responding to another bot. Bot pings off of two instances of itself with a 120 second delay without being unable to reply to users during that time--is the idea